% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{Factorial_loop}
\alias{Factorial_loop}
\title{Factorial con loop}
\usage{
Factorial_loop(n)
}
\arguments{
\item{n}{este es el numero}
}
\description{
Esto me permite saber el factorial
}
\examples{
Factorial_loop(20)


}
