% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{Factorial_reduce}
\alias{Factorial_reduce}
\title{Factorial}
\usage{
Factorial_reduce(n)
}
\arguments{
\item{n}{este es el numero}
}
\description{
Esto me permite saber el factorial
}
\examples{
Factorial_reduce(20)

}
