% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{is_natural}
\alias{is_natural}
\title{Es una funcion?}
\usage{
is_natural(n)
}
\arguments{
\item{n}{este es el numero}
}
\description{
Esto me permite saber si es un numero natural
}
\examples{
is_natural(20)

}
